requirements.txt
flask
flask-cors
openai
requests
python-docx
reportlab
